Add RunT using a minimal subset interface of *testing.T
dktest requires a `*testing.T` which limits where it can be used to
tests. There are some cases where a `*testing.T` is not available in
tests, where it's still useful to start a container, examples:

 * In a benchmark, we have a `*testing.B`, yet we may want to start a
   container with some workload that is being benchmarked.
 * When a lot of tests want to share a container, `TestMain` can be used
   to start a container, and have all tests use that container.

One workaround is for callers to pass a dummy value like `&testing.T{}`,
but this can result in less visibility as logs/fatals can be lost
silently.

Instead, use a minimal interface that is a subset of `testing.TB`, so
`*testing.T` and `*testing.B` can be passed, as well as a custom
implementation for other cases like `TestMain`.

Note: This adds a new method instead of updating the existing method, as
existing test functions are written against a `*testing.T`, and would
need to be updated.

Couple of other alternatives:
 * Use `testing.TB` as the interface instead of a new subset. This
   interface is much larger, so it's harder to provide a custom
   implementation in `TestMain`, but works for the benchmark case.
 * Add a separate method that doesn't require a testing.T or similar
   interface at all, only a logger, and it returns errors for Fatal.